feat: switch to update method
